requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
asyncpg==0.26.0
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
databases==0.6.0
ecdsa==0.18.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-decouple==3.6
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
websockets==10.3
